fixed error

+++ DataTypes.java
@@ -2,8 +2,7 @@ public class DataTypes
 {
     public static void run() {
         int i;
-        float f = 2.5;
+        double f = 2.5;
         boolean flag = true;
-        
     }
 }
